realtime/requirements.txt
Django>=5.0
channels
channels-redis
daphne
djangorestframework
djangorestframework-simplejwt
PyJWT
psycopg2-binary
python-dotenv
django-cors-headers
